Divide model.rb into each file
