+++ db/problem_group.rb
@@ -1,0 +1,40 @@
+class ProblemGroup < ActiveRecord::Base
+  validates :name,     presence: true
+
+  has_many :problems, dependent: :nullify
+
+  # method: POST
+  def self.allowed_to_create_by?(user = nil, action: "")
+    case user&.role_id
+    when ROLE_ID[:admin], ROLE_ID[:writer]
+      true
+    else # nologin, ...
+      false
+    end
+  end
+
+  # method: GET, PUT, PATCH, DELETE
+  def allowed?(method:, by: nil, action: "")
+    return self.class.readables(user: by, action: action).exists?(id: id) if method == "GET"
+
+    case by&.role_id
+    when ROLE_ID[:admin], ROLE_ID[:writer]
+      true
+    else # nologin, ...
+      false
+    end
+  end
+
+  # method: GET
+  scope :readables, ->(user: nil, action: "") {
+    case user&.role_id
+    when ROLE_ID[:admin], ROLE_ID[:writer], ROLE_ID[:viewer]
+      all
+    when ROLE_ID[:participant]
+      next none if DateTime.now <= Setting.competition_start_at
+      all
+    else # nologin, ...
+      none
+    end
+  }
+end
